fixes and graphql

+++ EindopdrachtBD/Program.cs
@@ -1,5 +1,4 @@
 var builder = WebApplication.CreateBuilder(args);
-
 //Database
 var dbSettings = builder.Configuration.GetSection("MongoConnection");
 builder.Services.Configure<DatabaseSettings>(dbSettings);
@@ -7,16 +6,40 @@ builder.Services.AddTransient<IMongoContext, MongoContext>();
 //Repositories
 builder.Services.AddTransient<IRestaurantRepository, RestaurantRepository>();
 builder.Services.AddTransient<IReviewRepository, ReviewRepository>();
+builder.Services.AddTransient<IUserRepository, UserRepository>();
 //Services
 builder.Services.AddTransient<IRestaurantService, RestaurantService>();
 builder.Services.AddTransient<IReviewService, ReviewService>();
+builder.Services.AddTransient<IAuthenticationService, AuthenticationService>();
 //Swagger
 builder.Services.AddSwaggerGen();
 builder.Services.AddEndpointsApiExplorer();
+//JWT authentication token
+var authsettings = builder.Configuration.GetSection("AuthenticationSettings");
+builder.Services.Configure<AuthenticationSettings>(authsettings);
+builder.Services.AddAuthorization(options =>{});
+builder.Services.AddAuthentication("Bearer").AddJwtBearer(options=>
+{
+    options.TokenValidationParameters = new(){
+        ValidateIssuer = true,
+        ValidateAudience = true,
+        ValidateIssuerSigningKey = true,
+        ValidIssuer = builder.Configuration["AuthenticationSettings:Issuer"],
+        ValidAudience = builder.Configuration["AuthenticationSettings:Audience"],
+        IssuerSigningKey = new SymmetricSecurityKey(System.Text.Encoding.ASCII.GetBytes(builder.Configuration["AuthenticationSettings:SecretForKey"]))
+    };
+}
+);
+//GraphQl
+builder.Services
+    .AddGraphQLServer()
+    .AddQueryType<Queries>()
+    .ModifyRequestOptions(opt => opt.IncludeExceptionDetails = true)
+    .AddMutationType<Mutation>();
+
 
 var app = builder.Build();
-
-//Add swagger documentation
+//Swagger documentation
 if (app.Environment.IsDevelopment())
 {
     app.UseSwagger();
@@ -26,13 +49,27 @@ if (app.Environment.IsDevelopment())
         options.RoutePrefix = string.Empty;
     });
 }
+//JWT authentication token
+app.UseAuthentication();
+app.UseAuthorization();
 
-app.MapGet("/", () => "Hello World!");
+
+//Setup
+app.MapGet("/setup", async (IRestaurantService restaurantService, IAuthenticationService authenticationService) => {
+    await authenticationService.Setup();
+    return Results.Created($"/restaurants", await restaurantService.Setup());
+});
+
+
+//Authentication/login
+app.MapPost("/authenticate", async (IAuthenticationService authenticationservice, AuthenticationRequestBody auth) => Results.Ok(await authenticationservice.Authenticate(auth)));
+
+
+//GraphQl
+app.MapGraphQL();
 
 
 //Restaurants
-app.MapGet("/setup", async (IRestaurantService restaurantService) => Results.Created($"/restaurants", await restaurantService.Setup()));
-
 app.MapGet("/restaurants", async (IRestaurantService restaurantService) => Results.Ok(await restaurantService.GetRestaurants()));
 
 app.MapGet("/restaurant/{id}", async (IRestaurantService restaurantService, string id) => Results.Ok(await restaurantService.GetRestaurant(id)));
@@ -42,15 +79,17 @@ app.MapPost("/restaurant", async (IRestaurantService restaurantService, Restaura
     return Results.Created($"/restaurant/{restaurant.RestaurantId}", results);
 });
 
-app.MapPut("/restaurant", async (IRestaurantService restaurantService, Restaurant restaurant) => Results.Ok(await restaurantService.UpdateRestaurant(restaurant)));
+app.MapPut("/restaurant", [Authorize] async (IRestaurantService restaurantService, Restaurant restaurant) => Results.Ok(await restaurantService.UpdateRestaurant(restaurant)));
 
-app.MapDelete("/restaurant/{id}", async (IRestaurantService restaurantService, string id) => Results.Ok(await restaurantService.DeleteRestaurant(id)));
+app.MapDelete("/restaurant/{id}", [Authorize] async (IRestaurantService restaurantService, string id) => Results.Ok(await restaurantService.DeleteRestaurant(id)));
 
 
 //Reviews
 app.MapGet("/reviews", async (IReviewService reviewService) => Results.Ok(await reviewService.GetReviews()));
 
 app.MapGet("/review/{id}", async (IReviewService reviewService, string id) => Results.Ok(await reviewService.GetReview(id)));
+
+app.MapGet("/restaurantreview/{id}", async (IReviewService reviewService, string id) => Results.Ok(await reviewService.GetReviewsRestaurant(id)));
 
 app.MapPost("/review", async (IReviewService reviewService, Review review) => {
     var results = await reviewService.AddReview(review);
@@ -61,8 +100,12 @@ app.MapPut("/review", async (IReviewService reviewService, Review review) => Res
 
 app.MapDelete("/review/{id}", async (IReviewService reviewService, string id) => Results.Ok(await reviewService.DeleteReview(id)));
 
+//Local/development
+//app.Run("http://localhost:3000");
 
-app.Run("http://localhost:3000");
-//app.Run();
-//public partial class Program { }
+//Docker
+app.Run();
+public partial class Program { }
+
+
 //app.Run("http://0.0.0:3000");
